Set ID on anchors
When reloading a page with a previous tab selected, the same tab is not setup correctly. The module currently expects the anchors to have an ID with a prefix of `label_`. Set this ID explicitly on each anchor so that it does not need to be included in the static markup.